Provide md-toc plugin for pre-commit framework
If a person uses https://pre-commit.com/ to configure tests
for their git repos, they can now use md-toc as a plugin.

This commit:

- Adds `.pre-commit-hooks.yaml` to define the hook.
  It is a simple definition because md-toc `setup.py`
  is well-defined and provides a default entrypoint.

- Updates `README.rst` to provide end-user instructions.

I was hesitant to add a test to the Makefile.
However, it is easy to test via:

    # all markdown files
    pre-commit try-repo . md-toc --all-files

    # or a specific file
    pre-commit try-repo . md-toc --file bar.md

If all `*.md` files that have the TOC marker are up-to-date,
the output looks like:

    Update markdown table-of-contents........................................Passed

If any `*.md` files that have the TOC marker are updated by the plugin,
pre-commit fails the run and warns that it modified 1 or more files:

    Update markdown table-of-contents........................................Failed
    - hook id: md-toc
    - files were modified by this hook

Pre-commit automatically fails if any file is modified, which
forces the user to add the updated file to the commit.

:bulb: If you merge this, it would help users if you bump the
minor tag. Pre-commit users usually put a `rev: <tag>` in their
`.pre-commit-config.yaml`. They _can_ put a git hash or
`rev: latest`, but a tag is much more common.

If you merge this, I am happy to then submit a PR to pre-commit to
have it included in the index at https://pre-commit.com/hooks.html